+++ src/layers/SplashesListPopup.h
@@ -5,7 +5,6 @@ using namespace geode::prelude;
 
 class SplashesListPopup : public geode::Popup<ArrayListNode*> {
     protected:
-        const cocos2d::CCPoint offset = {435.f / 2, 300.f / 2};
         ArrayListNode* m_node;
         CCMenuItemSpriteExtra* m_resetBtn;
         virtual bool setup(ArrayListNode* node) override;
@@ -16,6 +15,7 @@ class SplashesListPopup : public geode::Popup<ArrayListNode*> {
         virtual void deleteAllSplashes(cocos2d::CCObject* sender);
         virtual void resetSplashes(cocos2d::CCObject* sender);
     public:
+        const cocos2d::CCPoint offset = {435.f / 2, 300.f / 2};
         virtual void checkForChanges();
         virtual void updateSplashesList(float pos_x, float pos_y, float scale_x, float scale_y);
         static SplashesListPopup* create(ArrayListNode* node);